requirements.txt
flask==2.3.3
scikit-learn
pandas==2.0.3
requests==2.31.0
joblib==1.3.2
gunicorn
python-dotenv
flask-cors
google-generativeai